BACKEND/requirements.txt
Django==5.1.2
pillow==10.4.0
djangorestframework==3.15.2
django-jazzmin==3.0.1
django-cors-headers==4.5.0
drf-yasg==1.21.7